Hoist CapturePipe to shared helpers; rename GetWriter
- Rename Reporter::GetOutputWriter to GetWriter for brevity; matches
  the level-targeted call sites that follow it.
- Move CapturePipe and CaptureReporter from WSLCCLIReporterUnitTests.cpp
  into WSLCCLITestHelpers.h so other CLI unit test suites can capture
  Reporter output without duplicating the pipe plumbing.
- Fix the read-pipe handle leak in CapturePipe: hold the read end in a
  wil::unique_hfile member (declared before m_reader so destruction order
  joins the reader thread first, then closes the handle).
- Open the read pipe with ReadPipeOverlapped=true so PartialHandleRead's
  InterruptableRead can be interrupted by its exit event during teardown
  rather than relying on writer-side fclose ordering.
