+++ test/windows/wslc/WSLCCLIReporterUnitTests.cpp
@@ -28,89 +28,6 @@ using namespace WEX::TestExecution;
 
 namespace WSLCCLIReporterUnitTests {
 
-// RAII pipe pair for capturing FILE* output in tests.
-// write() end is passed to OutputChannel; captured() drains the read end after flush.
-struct CapturePipe
-{
-    CapturePipe()
-    {
-        auto [r, w] = wsl::windows::common::wslutil::OpenAnonymousPipe(0, false, false);
-        const int writeFd = _open_osfhandle(reinterpret_cast<intptr_t>(w.get()), _O_WRONLY | _O_TEXT);
-        THROW_HR_IF(E_FAIL, writeFd < 0);
-        w.release();
-
-        auto closeFd = wil::scope_exit([&] { _close(writeFd); });
-
-        WI_VERIFY(_setmode(writeFd, _O_U8TEXT) != -1);
-
-        m_file = _fdopen(writeFd, "w");
-        THROW_HR_IF(E_FAIL, m_file == nullptr);
-
-        closeFd.release();
-
-        m_reader = std::make_unique<PartialHandleRead>(r.release());
-    }
-
-    ~CapturePipe()
-    {
-        if (m_file)
-        {
-            fclose(m_file);
-        }
-    }
-
-    NON_COPYABLE(CapturePipe);
-    NON_MOVABLE(CapturePipe);
-
-    FILE* file() const
-    {
-        return m_file;
-    }
-
-    std::wstring captured()
-    {
-        if (m_file)
-        {
-            fclose(m_file);
-            m_file = nullptr;
-        }
-
-        m_reader->ExpectClosed();
-        std::wstring result = wsl::shared::string::MultiByteToWide(m_reader->GetData());
-
-        // _O_U8TEXT prepends a UTF-8 BOM on some streams; strip it if present.
-        if (!result.empty() && result[0] == L'\xFEFF')
-        {
-            result.erase(0, 1);
-        }
-
-        // _O_U8TEXT translates \n to \r\n; strip \r so tests compare plain newlines.
-        result.erase(std::remove(result.begin(), result.end(), L'\r'), result.end());
-        return result;
-    }
-
-private:
-    FILE* m_file = nullptr;
-    std::unique_ptr<PartialHandleRead> m_reader;
-};
-
-// Reporter wired to a single capture pipe for full output capture.
-// VT is disabled (not a console handle), so error output stays in the same pipe.
-struct CaptureReporter
-{
-    CapturePipe pipe;
-    Reporter reporter;
-
-    explicit CaptureReporter(bool vtEnabled = false) : reporter(pipe.file(), vtEnabled)
-    {
-    }
-
-    std::wstring captured()
-    {
-        return pipe.captured();
-    }
-};
-
 struct SplitCaptureReporter
 {
     CapturePipe outPipe;
